backend/pyproject.toml
[project]
name = "vulcan-chatbot"
version = "0.1.0"
description = "Web search chatbot with 2-Agent architecture"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "openai>=1.0.0",
    "anthropic>=0.30.0",
    "httpx>=0.27.0",
    "sse-starlette>=2.0.0",
    "python-dotenv>=1.0.0",
    "python-telegram-bot[ext]>=21.0",
    "apscheduler>=3.10.0",
    "aiosqlite>=0.20.0",
    "fugle-marketdata>=1.0.0",
    "finnhub-python>=2.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "respx>=0.21.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
